cleanup bounds definitions
