+++ src/dsl/object_types/deserialization.rs
@@ -1,9 +1,7 @@
 use crate::dsl::object_types::bounds::deserialization::deserialize_enumeration_values;
-use crate::dsl::object_types::bounds::IntegerBound;
-use crate::dsl::object_types::bounds::IntegerObjectBounds;
+use crate::dsl::object_types::bounds::deserialization::deserialize_integer_bounds;
 use crate::dsl::object_types::ObjectType;
 use crate::dsl::object_types::RawObjectType;
-use heck::MixedCase;
 use serde::de::Error;
 use serde_yaml::Mapping;
 use serde_yaml::Value;
@@ -27,62 +25,6 @@ where
     })?)
 }
 
-fn deserialize_integer<E>(name: &str, mapping: &Mapping) -> Result<Option<i64>, E>
-where
-    E: Error,
-{
-    let maximum_key = Value::from(name);
-    let value = mapping.get(&maximum_key);
-    println!("value from mapping {:#?} for {}", &value, &name);
-
-    match value {
-        None => Ok(None),
-        Some(value) => match value.as_i64() {
-            None => Err(Error::custom(format!("cannot deserialize {:#?} as integer", value))),
-            Some(value) => Ok(Some(value)),
-        },
-    }
-}
-
-fn deserialize_minmax<E>(name: &str, mapping: &Mapping) -> Result<Option<IntegerBound>, E>
-where
-    E: Error,
-{
-    let normal = deserialize_integer(name, mapping)?;
-    let exclusive = deserialize_integer(&("exclusive ".to_string() + name).to_mixed_case(), mapping)?;
-    if normal.is_some() && exclusive.is_some() {
-        return Err(Error::custom("cannot have both {} and exclusive {} set"));
-    }
-    if normal.is_some() {
-        return Ok(Some(IntegerBound::Inclusive(normal.unwrap())));
-    }
-    if exclusive.is_some() {
-        return Ok(Some(IntegerBound::Exclusive(exclusive.unwrap())));
-    }
-    Ok(None)
-}
-
-fn deserialize_integer_bounds<E>(mapping: &Mapping) -> Result<Option<IntegerObjectBounds>, E>
-where
-    E: Error,
-{
-    let maximum = deserialize_minmax("maximum", mapping)?;
-    let minimum = deserialize_minmax("minimum", mapping)?;
-    let multiple_of = deserialize_integer("multipleOf", mapping)?;
-    println!("maximum: {:#?}", &maximum);
-    println!("minimum: {:#?}", &minimum);
-
-    if maximum.is_some() {
-        Ok(Some(IntegerObjectBounds {
-            minimum,
-            maximum,
-            multiple_of,
-        }))
-    } else {
-        Ok(None)
-    }
-}
-
 impl RawObjectType {
     fn from<E>(value: &str, mapping: &Mapping) -> Result<Self, E>
     where
@@ -102,5 +44,22 @@ impl RawObjectType {
             _ => return Err(Error::custom(format!("unknown object type `{}`", value))),
         };
         Ok(object_type)
+    }
+}
+
+pub fn deserialize_integer<E>(name: &str, mapping: &Mapping) -> Result<Option<i64>, E>
+where
+    E: Error,
+{
+    let maximum_key = Value::from(name);
+    let value = mapping.get(&maximum_key);
+    println!("value from mapping {:#?} for {}", &value, &name);
+
+    match value {
+        None => Ok(None),
+        Some(value) => match value.as_i64() {
+            None => Err(Error::custom(format!("cannot deserialize {:#?} as integer", value))),
+            Some(value) => Ok(Some(value)),
+        },
     }
 }
